\name{DWD}
\docType{data}
\alias{DWD}
\title{Phenological observations}
\description{
	Phenological observations of nine stations from 1951 to 1998.
	Data from the German Weather Service.
}
\usage{data(Pheno)}
\format{Data frame containing three columns (day of year of observations,year,station-id)}
\source{German Weather Service}
\reference{
		Schaber J, Badeck F-W (2002) 'Evaluation of methods for the combination of phenological time series and outlier detection'. Tree Physiology 22:973-982
}
\keyword{datasets}
