\name{leapyear}
\alias{leapyear}
\title{Boolean test for leap year}
\description{
	Tests whether a given year is a leap year or not.
}
\usage{leapyear(y)}
\arguments{
	\item{y}{Year, numeric coerced into integer.}
}
\value{
	\item{TRUE}{leap year}	
	\item{FALSE}{non leap year}	
}
\author{
	Joerg Schaber
}
\examples{
	leapyear(2000)
	leapyear(2004)
}
\keyword{misc}
\keyword{utilities}
