\name{maxdaylength}
\alias{maxdaylength}
\title{Maximal day length on latitude l}
\description{
	Calculates maximal daylength maxdl [h] at a certain latitude l [degrees].	
}
\usage{maxdaylength(l)}
\arguments{
	\item{l}{Latitude in degrees.}
}
\value{
	\item{maxdl}{Maximal daylength [h] at a certain latitude l [degrees]}	
}
\author{
	Joerg Schaber
}
\examples{
	maxdaylength(60)
}
\keyword{misc}
\keyword{utilities}
