\name{Searle}
\docType{data}
\alias{Searle}
\title{Example of a two-way classification table}
\description{
	Example of a two-way classification table where lacking data
	creates three distinct connected sets.
}
\usage{data(Searle)}
\format{R source file}
\references{
	Searle (1997) 'Linear Models'. Wiley. 324p.
}	
\keyword{datasets}
